cdk_python/requirements.txt
aws-cdk-lib==2.165.0
constructs>=10.0.0,<11.0.0
boto3